Task 7/requirements.txt
streamlit
requests
Pillow
textblob
python-dotenv
langdetect
deep-translator
